+++ test_scripts/Polices/External_UCS/External_Consent_ON/008_ATF_Policies_External_Consent_ON_disallowed_by_external_consent_entities_omitted_user_consent_allowed.lua
@@ -1,8 +1,11 @@
+-------------------------------------- Requirement summary -------------------------------------------
+-- [Policies] External UCS: "ON" in case of omitted "disallowed_by_external_consent_entities" param in functional groupings
+--
 ------------------------------------------------------------------------------------------------------
 ------------------------------------General Settings for Configuration--------------------------------
 ------------------------------------------------------------------------------------------------------
 require('user_modules/all_common_modules')
-local disallowed_by_external_consent_entities_off = require('user_modules/shared_testcases_custom/ATF_Policies_External_Consent_common_functions')
+local common_functions_external_consent = require('user_modules/shared_testcases_custom/ATF_Policies_External_Consent_common_functions')
 ------------------------------------------------------------------------------------------------------
 ---------------------------------------Common Variables-----------------------------------------------
 ------------------------------------------------------------------------------------------------------
@@ -12,146 +15,139 @@ local policy_file = config.pathToSDL .. "storage/policy.sqlite"
 ---------------------------------------Preconditions--------------------------------------------------
 ------------------------------------------------------------------------------------------------------
 -- Start SDL and register application
-disallowed_by_external_consent_entities_off:PreconditonSteps("mobileConnection","mobileSession")
+common_functions_external_consent:PreconditonSteps("mobileConnection","mobileSession")
 -- Activate application
 common_steps:ActivateApplication("Activate_Application_1", config.application1.registerAppInterfaceParams.appName)
 ------------------------------------------------------------------------------------------------------
 ------------------------------------------Tests-------------------------------------------------------
 ------------------------------------------------------------------------------------------------------
 --------------------------------------------------------------------------
--- TEST 06:
-  -- In case
-  -- "functional grouping" is user_allowed by External Consent "ON" notification from HMI
-  -- and SDL gets SDL.OnAppPermissionConsent ( "functional grouping": userDisallowed, appID)from HMI
-  -- SDL must
-  -- update "consent_groups" of specific app (change appropriate <functional_grouping> status to "false")
-  -- leave the same value in "external_consent_status_groups" (<functional_grouping>:true)
-  -- send OnPermissionsChange to all impacted apps
-  -- send 'USER_DISALLOWED, success:false' to mobile app on requested RPCs from this "functional grouping"
+-- TEST 01:
+-- In case:
+-- SDL Policies database omits "disallowed_by_external_consent_entities_off" or "disallowed_by_external_consent_entities_on" param in "functional grouping" section
+-- and SDL gets SDL.OnAppPermissionConsent ("externalConsentStatus: ON")
+-- and registered application has this "functional grouping" assigned,
+-- SDL must process requested RPCs from "functional groupings" assigned to mobile app in terms of user_consent and data_consent policies rules
 --------------------------------------------------------------------------
--- Test 06.02:
--- Description:
---   "functional grouping" is allowed by External Consent "ON"
---     (disallowed_by_external_consent_entities_off exists. HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON))
---   HMI -> SDL: OnAppPermissionConsent(function = disallowed)
--- Expected Result:
---   Update: "consent_group"'s is_consented = 0.
---   Not update: "external_consent_status_groups" is_consented = 1.
---   OnPermissionsChange is sent.
---   Process RPCs from such "<functional_grouping>" as user disallowed
+-- Test 01.01:
+-- Description: disallowed_by_external_consent_entities_on/off is omitted. HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON, function allowed)
+-- Expected Result: requested RPC is allowed by user consent
 --------------------------------------------------------------------------
 -- Precondition:
---   Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
---   Request Policy Table Update.
+-- Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
+-- Request Policy Table Update.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON.."Precondition_Update_Policy_Table"] = function(self)
+Test[TEST_NAME_ON .. "Precondition_Update_Policy_Table"] = function(self)
   -- create json for PTU from sdl_preloaded_pt.json
-  local data = disallowed_by_external_consent_entities_off:ConvertPreloadedToJson()
+  local data = common_functions_external_consent:ConvertPreloadedToJson()
   -- insert Group001 into "functional_groupings"
   data.policy_table.functional_groupings.Group001 = {
     user_consent_prompt = "ConsentGroup001",
-    disallowed_by_external_consent_entities_off = {{
-      entityType = 1,
-      entityID = 1
-    }},
+    -- omit disallowed_by_external_consent_entities_on/off
     rpcs = {
       SubscribeWayPoints = {
         hmi_levels = {"BACKGROUND", "FULL", "LIMITED"}
       }
     }
   }
-  --insert application "0000001" which belong to functional group "Group001" into "app_policies"
+  -- insert Group002 into "functional_groupings"
+  data.policy_table.functional_groupings.Group002 = {
+    user_consent_prompt = "ConsentGroup002",
+    disallowed_by_external_consent_entities_off = {{
+        entityType = 2,
+        entityID = 5
+    }},
+    rpcs = {
+      SubscribeVehicleData = {
+        hmi_levels = {"BACKGROUND", "FULL", "LIMITED"}
+      }
+    }
+  }
+  --insert application "0000001" into "app_policies"
   data.policy_table.app_policies["0000001"] = {
     keep_context = false,
     steal_focus = false,
     priority = "NONE",
     default_hmi = "NONE",
-    groups = {"Base-4", "Group001"}
+    groups = {"Base-4", "Group001", "Group002"}
   }
   --insert "ConsentGroup001" into "consumer_friendly_messages"
   data.policy_table.consumer_friendly_messages.messages["ConsentGroup001"] = {languages = {}}
   data.policy_table.consumer_friendly_messages.messages.ConsentGroup001.languages["en-us"] = {
-        tts = "tts_test",
-        label = "label_test",
-        textBody = "textBody_test"
+    tts = "tts_test",
+    label = "label_test",
+    textBody = "textBody_test"
+  }
+  --insert "ConsentGroup002" into "consumer_friendly_messages"
+  data.policy_table.consumer_friendly_messages.messages["ConsentGroup002"] = {languages = {}}
+  data.policy_table.consumer_friendly_messages.messages.ConsentGroup002.languages["en-us"] = {
+    tts = "tts_test",
+    label = "label_test",
+    textBody = "textBody_test"
   }
   -- create json file for Policy Table Update
-  disallowed_by_external_consent_entities_off:CreateJsonFileForPTU(data, "/tmp/ptu_update.json")
+  common_functions_external_consent:CreateJsonFileForPTU(data, "/tmp/ptu_update.json")
   -- remove preload_pt from json file
   local parent_item = {"policy_table","module_config"}
   local removed_json_items = {"preloaded_pt"}
   common_functions:RemoveItemsFromJsonFile("/tmp/ptu_update.json", parent_item, removed_json_items)
   -- update policy table
-  disallowed_by_external_consent_entities_off:UpdatePolicy(self, "/tmp/ptu_update.json")
+  common_functions_external_consent:UpdatePolicy(self, "/tmp/ptu_update.json")
 end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   Check GetListOfPermissions response with empty externalConsentStatus array list. Get group id.
+-- Check GetListOfPermissions response with empty externalConsentStatus array list. Get group id.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON.."Precondition_GetListOfPermissions"] = function(self)
+Test[TEST_NAME_ON .. "Precondition_GetListOfPermissions"] = function(self)
   local request_id = self.hmiConnection:SendRequest("SDL.GetListOfPermissions")
   EXPECT_HMIRESPONSE(request_id,{
-    result = {
-      code = 0,
-      method = "SDL.GetListOfPermissions",
-      allowedFunctions = {{name = "ConsentGroup001", allowed = nil}},
-      externalConsentStatus = {}
-    }
-  })
+      result = {
+        code = 0,
+        method = "SDL.GetListOfPermissions",
+        allowedFunctions = {{name = "ConsentGroup001", allowed = nil}},
+        externalConsentStatus = {}
+      }
+    })
   :Do(function(_,data)
-    id_group_1 = disallowed_by_external_consent_entities_off:GetGroupId(data, "ConsentGroup001")
-  end)
+      id_group_1 = common_functions_external_consent:GetGroupId(data, "ConsentGroup001")
+    end)
 end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   HMI sends OnAppPermissionConsent with External Consent status = ON
---------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent"] = function(self)
-	self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
-    source = "GUI",
-    externalConsentStatus = {{entityType = 1, entityID = 1, status = "ON"}}
-  })
-  self.mobileSession:ExpectNotification("OnPermissionsChange")
-  :ValidIf(function(_,data)
-    local validate_result = disallowed_by_external_consent_entities_off:ValidateHMIPermissions(data,
-      "SubscribeWayPoints", {allowed = {"BACKGROUND","FULL","LIMITED"}, userDisallowed = {}})
-    return validate_result
-  end)
-end
-
---------------------------------------------------------------------------
--- Precondition:
---   OnAppPermissionChanged is not sent
---   when HMI sends OnAppPermissionConsent with consentedFunctions allowed = false
+-- HMI sends OnAppPermissionConsent with External Consent status = ON and consented function = allowed
 --------------------------------------------------------------------------
 Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent"] = function(self)
   hmi_app_id_1 = common_functions:GetHmiAppId(config.application1.registerAppInterfaceParams.appName, self)
-	self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
-    appID = hmi_app_id_1, source = "GUI",
-    consentedFunctions = {{name = "ConsentGroup001", id = id_group_1, allowed = false}}
-  })
+  self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
+      appID = hmi_app_id_1, source = "GUI",
+      externalConsentStatus = {{entityType = 2, entityID = 5, status = "ON"}},
+      consentedFunctions = {{name = "ConsentGroup001", id = id_group_1, allowed = true}}
+    })
   self.mobileSession:ExpectNotification("OnPermissionsChange")
   :ValidIf(function(_,data)
-    local validate_result = disallowed_by_external_consent_entities_off:ValidateHMIPermissions(data,
-      "SubscribeWayPoints", {allowed = {}, userDisallowed = {"BACKGROUND","FULL","LIMITED"}})
-    return validate_result
-  end)
+      local validate_result = common_functions_external_consent:ValidateHMIPermissions(data,
+        "SubscribeWayPoints", {allowed = {"BACKGROUND","FULL","LIMITED"}, userDisallowed = {}})
+      return validate_result
+    end)
 end
 
 --------------------------------------------------------------------------
 -- Main check:
---   RPC is disallowed to process.
+-- RPC is allowed to process.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "MainCheck_RPC_is_disallowed"] = function(self)
+Test[TEST_NAME_ON .. "MainCheck_RPC_is_allowed"] = function(self)
   local corid = self.mobileSession:SendRPC("SubscribeWayPoints",{})
-  EXPECT_RESPONSE("SubscribeWayPoints", {success = false , resultCode = "USER_DISALLOWED"})
+  EXPECT_HMICALL("Navigation.SubscribeWayPoints")
+  :Do(function(_,data)
+      self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
+    end)
+  EXPECT_RESPONSE("SubscribeWayPoints", {success = true , resultCode = "SUCCESS"})
   EXPECT_NOTIFICATION("OnHashChange")
-  :Times(0)
 end
 
--- end Test 06.02
+-- end Test 01.01
 ----------------------------------------------------
 ---------------------------------------------------------------------------------------------
 --------------------------------------Postcondition------------------------------------------
