+++ test_scripts/Polices/External_UCS/External_Consent_ON/014_ATF_Policies_External_Consent_ON_disallowed_change_to_OFF_allowed.lua
@@ -1,58 +1,60 @@
+-------------------------------------- Requirement summary -------------------------------------------
+-- [Policies] External UCS: "ON" updates in userDisallowed "consent_groups" and "external_consent_status_groups" when externalConsentStatus changes to "OFF"
+--
 ------------------------------------------------------------------------------------------------------
 ------------------------------------General Settings for Configuration--------------------------------
 ------------------------------------------------------------------------------------------------------
 require('user_modules/all_common_modules')
-local common_functions_external_consent = require('user_modules/shared_testcases_custom/ATF_Policies_External_Consent_common_functions')
+local disallowed_by_external_consent_entities_on = require('user_modules/shared_testcases_custom/ATF_Policies_External_Consent_common_functions')
 ------------------------------------------------------------------------------------------------------
 ---------------------------------------Common Variables-----------------------------------------------
 ------------------------------------------------------------------------------------------------------
-local id_group_1
 local policy_file = config.pathToSDL .. "storage/policy.sqlite"
 ------------------------------------------------------------------------------------------------------
 ---------------------------------------Preconditions--------------------------------------------------
 ------------------------------------------------------------------------------------------------------
 -- Start SDL and register application
-common_functions_external_consent:PreconditonSteps("mobileConnection","mobileSession")
+disallowed_by_external_consent_entities_on:PreconditonSteps("mobileConnection","mobileSession")
 -- Activate application
 common_steps:ActivateApplication("Activate_Application_1", config.application1.registerAppInterfaceParams.appName)
 ------------------------------------------------------------------------------------------------------
 ------------------------------------------Tests-------------------------------------------------------
 ------------------------------------------------------------------------------------------------------
 --------------------------------------------------------------------------
--- TEST 05:
-  -- In case
-  -- "functional grouping" is user_disallowed by External Consent "ON" notification from HMI
-  -- and SDL gets SDL.OnAppPermissionConsent ( "functional grouping": allowed, appID)from HMI
-  -- SDL must
-  -- update "consent_groups" of specific app (change appropriate <functional_grouping> status to "true")
-  -- leave the same value in "external_consent_status_groups" (<functional_grouping>:false)
-  -- send OnPermissionsChange to all impacted apps
-  -- process RPCs from such "<functional_grouping>" as user allowed
+-- TEST 08:
+-- In case
+-- SDL gets SDL.OnAppPermissionConsent with externalConsentStatus change from "ON" to "OFF"
+-- and this changes the "<functional grouping>" permissions to "user_allowed",
+-- SDL must
+-- for all impacted apps
+-- update "consent_groups" to status "true"
+-- update "external_consent_status_groups" to status "true"
+-- send OnPermissionsChange ("RPCs from assigned "functional_groupings" "hmiPermission": (<current assigned permissions>)", allowed) notification
 --------------------------------------------------------------------------
--- Test 05.02:
+-- Test 08.01:
 -- Description:
---   "functional grouping" is user_disallowed by External Consent "ON"
---     (disallowed_by_external_consent_entities_on exists. HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON))
---   HMI -> SDL: OnAppPermissionConsent(function = disallowed)
+-- "functional grouping" is user_disallowed by External Consent "ON"
+-- (disallowed_by_external_consent_entities_on exists. HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON))
+-- HMI -> SDL: OnAppPermissionConsent(externalConsentStatus OFF)
 -- Expected Result:
---   Not update: "consent_group"'s is_consented = 0.
---   Not update: "external_consent_status_groups" is_consented = 0.
---   OnPermissionsChange is not sent.
---   Process RPCs from such "<functional_grouping>" as user disallowed
+-- Update: "consent_group"'s is_consented = 1.
+-- Update: "external_consent_status_groups" is_consented = 1.
+-- OnPermissionsChange is sent.
+-- Process RPCs from such "<functional_grouping>" as user allowed
 --------------------------------------------------------------------------
 -- Precondition:
---   Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
---   Request Policy Table Update.
+-- Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
+-- Request Policy Table Update.
 --------------------------------------------------------------------------
 Test[TEST_NAME_ON.."Precondition_Update_Policy_Table"] = function(self)
   -- create json for PTU from sdl_preloaded_pt.json
-  local data = common_functions_external_consent:ConvertPreloadedToJson()
+  local data = disallowed_by_external_consent_entities_on:ConvertPreloadedToJson()
   -- insert Group001 into "functional_groupings"
   data.policy_table.functional_groupings.Group001 = {
     user_consent_prompt = "ConsentGroup001",
     disallowed_by_external_consent_entities_on = {{
-      entityType = 1,
-      entityID = 1
+        entityType = 2,
+        entityID = 5
     }},
     rpcs = {
       SubscribeWayPoints = {
@@ -71,80 +73,62 @@ Test[TEST_NAME_ON.."Precondition_Update_Policy_Table"] = function(self)
   --insert "ConsentGroup001" into "consumer_friendly_messages"
   data.policy_table.consumer_friendly_messages.messages["ConsentGroup001"] = {languages = {}}
   data.policy_table.consumer_friendly_messages.messages.ConsentGroup001.languages["en-us"] = {
-        tts = "tts_test",
-        label = "label_test",
-        textBody = "textBody_test"
+    tts = "tts_test",
+    label = "label_test",
+    textBody = "textBody_test"
   }
   -- create json file for Policy Table Update
-  common_functions_external_consent:CreateJsonFileForPTU(data, "/tmp/ptu_update.json")
+  disallowed_by_external_consent_entities_on:CreateJsonFileForPTU(data, "/tmp/ptu_update.json")
   -- remove preload_pt from json file
   local parent_item = {"policy_table","module_config"}
   local removed_json_items = {"preloaded_pt"}
   common_functions:RemoveItemsFromJsonFile("/tmp/ptu_update.json", parent_item, removed_json_items)
   -- update policy table
-  common_functions_external_consent:UpdatePolicy(self, "/tmp/ptu_update.json")
+  disallowed_by_external_consent_entities_on:UpdatePolicy(self, "/tmp/ptu_update.json")
 end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   Check GetListOfPermissions response with empty externalConsentStatus array list. Get group id.
+-- Check GetListOfPermissions response with empty externalConsentStatus array list. Get group id.
 --------------------------------------------------------------------------
 Test[TEST_NAME_ON.."Precondition_GetListOfPermissions"] = function(self)
   --hmi side: sending SDL.GetListOfPermissions request to SDL
   local request_id = self.hmiConnection:SendRequest("SDL.GetListOfPermissions")
   -- hmi side: expect SDL.GetListOfPermissions response
   EXPECT_HMIRESPONSE(request_id,{
-    result = {
-      code = 0,
-      method = "SDL.GetListOfPermissions",
-      allowedFunctions = {{name = "ConsentGroup001", allowed = nil}},
-      externalConsentStatus = {}
-    }
-  })
-  :Do(function(_,data)
-    id_group_1 = common_functions_external_consent:GetGroupId(data, "ConsentGroup001")
-  end)
+      result = {
+        code = 0,
+        method = "SDL.GetListOfPermissions",
+        allowedFunctions = {{name = "ConsentGroup001", allowed = nil}},
+        externalConsentStatus = {}
+      }
+    })
 end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   HMI sends OnAppPermissionConsent with External Consent status = ON
+-- HMI sends OnAppPermissionConsent with External Consent status = ON
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent"] = function(self)
+Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent_externalConsentStatus_ON"] = function(self)
   -- hmi side: sending SDL.OnAppPermissionConsent for applications
-	self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
-    source = "GUI",
-    externalConsentStatus = {{entityType = 1, entityID = 1, status = "ON"}}
-  })
+  self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
+      source = "GUI",
+      externalConsentStatus = {{entityType = 2, entityID = 5, status = "ON"}}
+    })
   self.mobileSession:ExpectNotification("OnPermissionsChange")
   :ValidIf(function(_,data)
-    local validate_result = common_functions_external_consent:ValidateHMIPermissions(data,
-      "SubscribeWayPoints", {allowed = {}, userDisallowed = {"BACKGROUND","FULL","LIMITED"}})
-    return validate_result
-  end)
+      local validate_result = disallowed_by_external_consent_entities_on:ValidateHMIPermissions(data,
+        "SubscribeWayPoints", {allowed = {}, userDisallowed = {"BACKGROUND","FULL","LIMITED"}})
+      return validate_result
+    end)
 end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   OnAppPermissionChanged is not sent
---   when HMI sends OnAppPermissionConsent without consentedFunctions
+-- RPC is disallowed to process.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent"] = function(self)
-  hmi_app_id_1 = common_functions:GetHmiAppId(config.application1.registerAppInterfaceParams.appName, self)
-  -- hmi side: sending SDL.OnAppPermissionConsent for applications
-	self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
-    appID = hmi_app_id_1, source = "GUI"
-  })
-  self.mobileSession:ExpectNotification("OnPermissionsChange")
-  :Times(0)
-end
-
---------------------------------------------------------------------------
--- Main check:
---   RPC is disallowed to process.
---------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "MainCheck_RPC_is_disallowed"] = function(self)
-	--mobile side: send SubscribeWayPoints request
+Test[TEST_NAME_ON .. "Precondition_RPC_is_disallowed"] = function(self)
+  --mobile side: send SubscribeWayPoints request
   local corid = self.mobileSession:SendRPC("SubscribeWayPoints",{})
   --mobile side: SubscribeWayPoints response
   EXPECT_RESPONSE("SubscribeWayPoints", {success = false , resultCode = "USER_DISALLOWED"})
@@ -152,7 +136,45 @@ Test[TEST_NAME_ON .. "MainCheck_RPC_is_disallowed"] = function(self)
   :Times(0)
 end
 
--- end Test 05.02
+--------------------------------------------------------------------------
+-- Main check:
+-- OnAppPermissionChanged is sent
+-- when HMI sends OnAppPermissionConsent with externalConsentStatus = OFF
+--------------------------------------------------------------------------
+Test[TEST_NAME_ON .. "MainCheck_HMI_sends_OnAppPermissionConsent_externalConsentStatus_OFF"] = function(self)
+  hmi_app_id_1 = common_functions:GetHmiAppId(config.application1.registerAppInterfaceParams.appName, self)
+  -- hmi side: sending SDL.OnAppPermissionConsent for applications
+  self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
+      source = "GUI",
+      externalConsentStatus = {{entityType = 2, entityID = 5, status = "OFF"}}
+    })
+  self.mobileSession:ExpectNotification("OnPermissionsChange")
+  :ValidIf(function(_,data)
+      local validate_result = disallowed_by_external_consent_entities_on:ValidateHMIPermissions(data,
+        "SubscribeWayPoints", {allowed = {"BACKGROUND","FULL","LIMITED"}, userDisallowed = {}})
+      return validate_result
+    end)
+end
+
+--------------------------------------------------------------------------
+-- Main check:
+-- RPC is allowed to process.
+--------------------------------------------------------------------------
+Test[TEST_NAME_ON .. "MainCheck_RPC_is_allowed"] = function(self)
+  --mobile side: send SubscribeWayPoints request
+  local corid = self.mobileSession:SendRPC("SubscribeWayPoints",{})
+  --hmi side: expected SubscribeWayPoints request
+  EXPECT_HMICALL("Navigation.SubscribeWayPoints")
+  :Do(function(_,data)
+      --hmi side: sending Navigation.SubscribeWayPoints response
+      self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
+    end)
+  --mobile side: SubscribeWayPoints response
+  EXPECT_RESPONSE("SubscribeWayPoints", {success = true , resultCode = "SUCCESS"})
+  EXPECT_NOTIFICATION("OnHashChange")
+end
+
+-- end Test 08.01
 ----------------------------------------------------
 ---------------------------------------------------------------------------------------------
 --------------------------------------Postcondition------------------------------------------
