+++ test_scripts/Polices/External_UCS/External_Consent_ON/018_ATF_Policies_External_Consent_ON_external_consent_allowed_user_consent_allowed.lua
@@ -1,3 +1,7 @@
+-------------------------------------- Requirement summary -------------------------------------------
+-- [Policies] External UCS: "ON" user disables "functional grouping" for specific app
+-- [Policies] External UCS: "ON" updates in "consent_groups" and "external_consent_status_groups" when user disables <functional_grouping> for specific app
+--
 ------------------------------------------------------------------------------------------------------
 ------------------------------------General Settings for Configuration--------------------------------
 ------------------------------------------------------------------------------------------------------
@@ -6,6 +10,7 @@ local common_functions_external_consent = require('user_modules/shared_testcases
 ------------------------------------------------------------------------------------------------------
 ---------------------------------------Common Variables-----------------------------------------------
 ------------------------------------------------------------------------------------------------------
+local id_group_1
 local policy_file = config.pathToSDL .. "storage/policy.sqlite"
 ------------------------------------------------------------------------------------------------------
 ---------------------------------------Preconditions--------------------------------------------------
@@ -18,34 +23,43 @@ common_steps:ActivateApplication("Activate_Application_1", config.application1.r
 ------------------------------------------Tests-------------------------------------------------------
 ------------------------------------------------------------------------------------------------------
 --------------------------------------------------------------------------
--- TEST:
-  -- externalConsentStatus is not applied when user_consent_prompt does not exist in functional group
+-- TEST 06:
+-- In case
+-- "functional grouping" is user_allowed by External Consent "ON" notification from HMI
+-- and SDL gets SDL.OnAppPermissionConsent ( "functional grouping": userDisallowed, appID)from HMI
+-- SDL must
+-- update "consent_groups" of specific app (change appropriate <functional_grouping> status to "false")
+-- leave the same value in "external_consent_status_groups" (<functional_grouping>:true)
+-- send OnPermissionsChange to all impacted apps
+-- send 'USER_DISALLOWED, success:false' to mobile app on requested RPCs from this "functional grouping"
 --------------------------------------------------------------------------
--- Test:
+-- Test 06.01:
 -- Description:
---   user_consent_prompt does not exist in functional group.  HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON))
+-- "functional grouping" is allowed by External Consent "ON"
+-- (disallowed_by_external_consent_entities_off exists. HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON))
+-- HMI -> SDL: OnAppPermissionConsent(function = allowed)
 -- Expected Result:
---   "consent_group" is not added.
---   "external_consent_status_groups" is not added.
---   OnPermissionsChange is not sent.
---   Process RPCs from such "<functional_grouping>" as user allowed
+-- Note update: "consent_group"'s is_consented = 1.
+-- Not update: "external_consent_status_groups" is_consented = 1.
+-- OnPermissionsChange is not sent.
+-- Process RPCs from such "<functional_grouping>" as user allowed
 --------------------------------------------------------------------------
 -- Precondition:
---   Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
---   Request Policy Table Update.
+-- Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
+-- Request Policy Table Update.
 --------------------------------------------------------------------------
 Test[TEST_NAME_ON.."Precondition_Update_Policy_Table"] = function(self)
   -- create json for PTU from sdl_preloaded_pt.json
   local data = common_functions_external_consent:ConvertPreloadedToJson()
   -- insert Group001 into "functional_groupings"
   data.policy_table.functional_groupings.Group001 = {
---    user_consent_prompt = "ConsentGroup001",
-    disallowed_by_external_consent_entities_on = {{
-      entityType = 2,
-      entityID = 5
+    user_consent_prompt = "ConsentGroup001",
+    disallowed_by_external_consent_entities_off = {{
+        entityType = 1,
+        entityID = 1
     }},
     rpcs = {
-      SendLocation = {
+      SubscribeWayPoints = {
         hmi_levels = {"BACKGROUND", "FULL", "LIMITED"}
       }
     }
@@ -61,9 +75,9 @@ Test[TEST_NAME_ON.."Precondition_Update_Policy_Table"] = function(self)
   --insert "ConsentGroup001" into "consumer_friendly_messages"
   data.policy_table.consumer_friendly_messages.messages["ConsentGroup001"] = {languages = {}}
   data.policy_table.consumer_friendly_messages.messages.ConsentGroup001.languages["en-us"] = {
-        tts = "tts_test",
-        label = "label_test",
-        textBody = "textBody_test"
+    tts = "tts_test",
+    label = "label_test",
+    textBody = "textBody_test"
   }
   -- create json file for Policy Table Update
   common_functions_external_consent:CreateJsonFileForPTU(data, "/tmp/ptu_update.json")
@@ -77,86 +91,78 @@ end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   Check GetListOfPermissions response with empty externalConsentStatus array list. Get group id.
+-- Check GetListOfPermissions response with empty externalConsentStatus array list. Get group id.
 --------------------------------------------------------------------------
 Test[TEST_NAME_ON.."Precondition_GetListOfPermissions"] = function(self)
   --hmi side: sending SDL.GetListOfPermissions request to SDL
   local request_id = self.hmiConnection:SendRequest("SDL.GetListOfPermissions")
   -- hmi side: expect SDL.GetListOfPermissions response
   EXPECT_HMIRESPONSE(request_id,{
-    result = {
-      code = 0,
-      method = "SDL.GetListOfPermissions",
-      allowedFunctions = {},
-      externalConsentStatus = {}
-    }
-  })
-end
-
---------------------------------------------------------------------------
--- Main Check:
---   OnAppPermissionChanged is NOT sent
---   when HMI sends OnAppPermissionConsent with externalConsentStatus = ON
---------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent_externalConsentStatus_ON"] = function(self)
-  -- hmi side: sending SDL.OnAppPermissionConsent for applications
-	self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
-    source = "GUI",
-    externalConsentStatus = {{entityType = 2, entityID = 5, status = "ON"}}
-  })
-  EXPECT_NOTIFICATION("OnPermissionsChange")
-  :Times(0)
-end
-
---------------------------------------------------------------------------
--- Main check:
---   RPC is allowed to process.
---------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "MainCheck_RPC_is_allowed_When_External_Consent_ON"] = function(self)
-  local cid = self.mobileSession:SendRPC("SendLocation", {
-    longitudeDegrees = 1.1,
-    latitudeDegrees = 1.1
-  })
-  EXPECT_HMICALL("Navigation.SendLocation")
+      result = {
+        code = 0,
+        method = "SDL.GetListOfPermissions",
+        allowedFunctions = {{name = "ConsentGroup001", allowed = nil}},
+        externalConsentStatus = {}
+      }
+    })
   :Do(function(_,data)
-    self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
-  end)
-  EXPECT_RESPONSE("SendLocation", {success = true , resultCode = "SUCCESS"})
+      id_group_1 = common_functions_external_consent:GetGroupId(data, "ConsentGroup001")
+    end)
 end
 
 --------------------------------------------------------------------------
--- Main check:
---   OnAppPermissionChanged is NOT sent
---   when HMI sends OnAppPermissionConsent with externalConsentStatus = OFF
+-- Precondition:
+-- HMI sends OnAppPermissionConsent with External Consent status = ON
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "MainCheck_HMI_sends_OnAppPermissionConsent_externalConsentStatus_OFF"] = function(self)
+Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent_externalConsentStatus"] = function(self)
+  -- hmi side: sending SDL.OnAppPermissionConsent for applications
+  self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
+      source = "GUI",
+      externalConsentStatus = {{entityType = 1, entityID = 1, status = "ON"}}
+    })
+  self.mobileSession:ExpectNotification("OnPermissionsChange")
+  :ValidIf(function(_,data)
+      local validate_result = common_functions_external_consent:ValidateHMIPermissions(data,
+        "SubscribeWayPoints", {allowed = {"BACKGROUND","FULL","LIMITED"}, userDisallowed = {}})
+      return validate_result
+    end)
+end
+
+--------------------------------------------------------------------------
+-- Precondition:
+-- OnAppPermissionChanged is sent
+-- when HMI sends OnAppPermissionConsent with consentedFunctions allowed = true
+--------------------------------------------------------------------------
+Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent_userConsent"] = function(self)
   hmi_app_id_1 = common_functions:GetHmiAppId(config.application1.registerAppInterfaceParams.appName, self)
   -- hmi side: sending SDL.OnAppPermissionConsent for applications
-	self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
-    source = "GUI",
-    externalConsentStatus = {{entityType = 2, entityID = 5, status = "OFF"}}
-  })
-  EXPECT_NOTIFICATION("OnPermissionsChange")
+  self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
+      appID = hmi_app_id_1, source = "GUI",
+      consentedFunctions = {{name = "ConsentGroup001", id = id_group_1, allowed = true}}
+    })
+  self.mobileSession:ExpectNotification("OnPermissionsChange")
   :Times(0)
 end
 
 --------------------------------------------------------------------------
 -- Main check:
---   RPC is allowed to process.
+-- RPC is allowed to process.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "MainCheck_RPC_is_allowed_When_External_Consent_OFF"] = function(self)
-  local cid = self.mobileSession:SendRPC("SendLocation", {
-    longitudeDegrees = 1.1,
-    latitudeDegrees = 1.1
-  })
-  EXPECT_HMICALL("Navigation.SendLocation")
+Test[TEST_NAME_ON .. "MainCheck_RPC_is_allowed"] = function(self)
+  --mobile side: send SubscribeWayPoints request
+  local corid = self.mobileSession:SendRPC("SubscribeWayPoints",{})
+  --hmi side: expected SubscribeWayPoints request
+  EXPECT_HMICALL("Navigation.SubscribeWayPoints")
   :Do(function(_,data)
-    self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
-  end)
-  EXPECT_RESPONSE("SendLocation", {success = true , resultCode = "SUCCESS"})
+      --hmi side: sending Navigation.SubscribeWayPoints response
+      self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
+    end)
+  --mobile side: SubscribeWayPoints response
+  EXPECT_RESPONSE("SubscribeWayPoints", {success = true , resultCode = "SUCCESS"})
+  EXPECT_NOTIFICATION("OnHashChange")
 end
 
--- end Test
+-- end Test 06.01
 ----------------------------------------------------
 ---------------------------------------------------------------------------------------------
 --------------------------------------Postcondition------------------------------------------
