+++ test_scripts/Polices/External_UCS/External_Consent_ON/022_ATF_Policies_External_Consent_ON_regist_new_app_disallowed_by_external_consent_entities_off.lua
@@ -1,3 +1,6 @@
+-------------------------------------- Requirement summary -------------------------------------------
+-- [Policies] External UCS: settings for all connected apps
+--
 ------------------------------------------------------------------------------------------------------
 ------------------------------------General Settings for Configuration--------------------------------
 ------------------------------------------------------------------------------------------------------
@@ -6,7 +9,6 @@ local common_functions_external_consent = require('user_modules/shared_testcases
 ------------------------------------------------------------------------------------------------------
 ---------------------------------------Common Variables-----------------------------------------------
 ------------------------------------------------------------------------------------------------------
-local id_group_1
 local policy_file = config.pathToSDL .. "storage/policy.sqlite"
 ------------------------------------------------------------------------------------------------------
 ---------------------------------------Preconditions--------------------------------------------------
@@ -16,52 +18,68 @@ common_functions_external_consent:PreconditonSteps("mobileConnection","mobileSes
 -- Activate application
 common_steps:ActivateApplication("Activate_Application_1", config.application1.registerAppInterfaceParams.appName)
 ------------------------------------------------------------------------------------------------------
+-------------------------------------------Functions--------------------------------------------------
+------------------------------------------------------------------------------------------------------
+local function RegistAndActivateApp(test_case_name, mobile_session, application_params)
+  common_steps:AddMobileSession("Precondition_Add_Mobile_Session_" .. test_case_name, "mobileConnection",mobile_session)
+  common_steps:RegisterApplication("Precondition_Register_Application_" .. test_case_name, mobile_session, application_params)
+  common_steps:ActivateApplication("Precondition_Activate_Application_" .. test_case_name, application_params.appName)
+end
+------------------------------------------------------------------------------------------------------
 ------------------------------------------Tests-------------------------------------------------------
 ------------------------------------------------------------------------------------------------------
 --------------------------------------------------------------------------
--- TEST 05:
-  -- In case
-  -- "functional grouping" is user_disallowed by External Consent "ON" notification from HMI
-  -- and SDL gets SDL.OnAppPermissionConsent ( "functional grouping": allowed, appID)from HMI
-  -- SDL must
-  -- update "consent_groups" of specific app (change appropriate <functional_grouping> status to "true")
-  -- leave the same value in "external_consent_status_groups" (<functional_grouping>:false)
-  -- send OnPermissionsChange to all impacted apps
-  -- process RPCs from such "<functional_grouping>" as user allowed
+-- TEST 12:
+-- In case
+-- SDL received SDL.OnAppPermissionConsent (externalConsentStatus) that changed user's permissions for a "<functional_grouping>"
+-- and new app that has this "<functional_grouping>" assigned connects
+-- SDL must
+-- 1. apply External Consent User Consent Settings to this app
+-- 2. send corresponding OnPermissionsChange
+-- 3. add corresponding records to PolicyTable -> "<deviceID>" section
 --------------------------------------------------------------------------
--- Test 05.01:
--- Description:
---   "functional grouping" is user_disallowed by External Consent "ON"
---     (disallowed_by_external_consent_entities_on exists. HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON))
---   HMI -> SDL: OnAppPermissionConsent(function = allowed)
--- Expected Result:
---   Update: "consent_group"'s is_consented = 1.
---   Not update: "external_consent_status_groups" is_consented = 0.
---   OnPermissionsChange is sent.
---   Process RPCs from such "<functional_grouping>" as user allowed
+-- Test 12.01:
+-- Description: disallowed_by_external_consent_entities_off. HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON). Register new applications.
+-- Expected Result: external_consent is created automatically.
 --------------------------------------------------------------------------
 -- Precondition:
---   Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
---   Request Policy Table Update.
+-- Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
+-- Request Policy Table Update.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON.."Precondition_Update_Policy_Table"] = function(self)
+Test[TEST_NAME_ON.."Precondition_Update_Policy_Table_1"] = function(self)
   -- create json for PTU from sdl_preloaded_pt.json
   local data = common_functions_external_consent:ConvertPreloadedToJson()
   -- insert Group001 into "functional_groupings"
   data.policy_table.functional_groupings.Group001 = {
     user_consent_prompt = "ConsentGroup001",
-    disallowed_by_external_consent_entities_on = {{
-      entityType = 1,
-      entityID = 1
+    disallowed_by_external_consent_entities_off = {{
+        entityType = 2,
+        entityID = 5
     }},
     rpcs = {
-      SubscribeWayPoints = {
+      SendLocation = {
         hmi_levels = {"BACKGROUND", "FULL", "LIMITED"}
       }
     }
   }
-  --insert application "0000001" which belong to functional group "Group001" into "app_policies"
+  --insert application "0000001" into "app_policies"
   data.policy_table.app_policies["0000001"] = {
+    keep_context = false,
+    steal_focus = false,
+    priority = "NONE",
+    default_hmi = "NONE",
+    groups = {"Base-4", "Group001"}
+  }
+  --insert application "0000002" into "app_policies"
+  data.policy_table.app_policies["0000002"] = {
+    keep_context = false,
+    steal_focus = false,
+    priority = "NONE",
+    default_hmi = "NONE",
+    groups = {"Base-4", "Group001"}
+  }
+  --insert application "0000003" into "app_policies"
+  data.policy_table.app_policies["0000003"] = {
     keep_context = false,
     steal_focus = false,
     priority = "NONE",
@@ -71,9 +89,9 @@ Test[TEST_NAME_ON.."Precondition_Update_Policy_Table"] = function(self)
   --insert "ConsentGroup001" into "consumer_friendly_messages"
   data.policy_table.consumer_friendly_messages.messages["ConsentGroup001"] = {languages = {}}
   data.policy_table.consumer_friendly_messages.messages.ConsentGroup001.languages["en-us"] = {
-        tts = "tts_test",
-        label = "label_test",
-        textBody = "textBody_test"
+    tts = "tts_test",
+    label = "label_test",
+    textBody = "textBody_test"
   }
   -- create json file for Policy Table Update
   common_functions_external_consent:CreateJsonFileForPTU(data, "/tmp/ptu_update.json")
@@ -87,82 +105,82 @@ end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   Check GetListOfPermissions response with empty externalConsentStatus array list. Get group id.
+-- Check GetListOfPermissions response with empty externalConsentStatus array list.
 --------------------------------------------------------------------------
 Test[TEST_NAME_ON.."Precondition_GetListOfPermissions"] = function(self)
   --hmi side: sending SDL.GetListOfPermissions request to SDL
   local request_id = self.hmiConnection:SendRequest("SDL.GetListOfPermissions")
   -- hmi side: expect SDL.GetListOfPermissions response
   EXPECT_HMIRESPONSE(request_id,{
-    result = {
-      code = 0,
-      method = "SDL.GetListOfPermissions",
-      allowedFunctions = {{name = "ConsentGroup001", allowed = nil}},
-      externalConsentStatus = {}
-    }
-  })
-  :Do(function(_,data)
-    id_group_1 = common_functions_external_consent:GetGroupId(data, "ConsentGroup001")
-  end)
+      result = {
+        code = 0,
+        method = "SDL.GetListOfPermissions",
+        allowedFunctions = {
+          {name = "ConsentGroup001", allowed = nil}
+        },
+        externalConsentStatus = {}
+      }
+    })
 end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   HMI sends OnAppPermissionConsent with External Consent status = ON
+-- HMI sends OnAppPermissionConsent with External Consent status = ON
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent_externalConsentStatus"] = function(self)
-  -- hmi side: sending SDL.OnAppPermissionConsent for applications
-	self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
-    source = "GUI",
-    externalConsentStatus = {{entityType = 1, entityID = 1, status = "ON"}}
-  })
-  self.mobileSession:ExpectNotification("OnPermissionsChange")
-  :ValidIf(function(_,data)
-    local validate_result = common_functions_external_consent:ValidateHMIPermissions(data,
-      "SubscribeWayPoints", {allowed = {}, userDisallowed = {"BACKGROUND","FULL","LIMITED"}})
-    return validate_result
-  end)
-end
-
---------------------------------------------------------------------------
--- Precondition:
---   OnAppPermissionChanged is sent
---   when HMI sends OnAppPermissionConsent with consentedFunctions allowed = true
---------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent_userConsent"] = function(self)
+Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent"] = function(self)
   hmi_app_id_1 = common_functions:GetHmiAppId(config.application1.registerAppInterfaceParams.appName, self)
   -- hmi side: sending SDL.OnAppPermissionConsent for applications
-	self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
-    appID = hmi_app_id_1, source = "GUI",
-    consentedFunctions = {{name = "ConsentGroup001", id = id_group_1, allowed = true}}
-  })
+  self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
+      appID = hmi_app_id_1, source = "GUI",
+      externalConsentStatus = {{entityType = 2, entityID = 5, status = "ON"}}
+    })
   self.mobileSession:ExpectNotification("OnPermissionsChange")
   :ValidIf(function(_,data)
-    local validate_result = common_functions_external_consent:ValidateHMIPermissions(data,
-      "SubscribeWayPoints", {allowed = {"BACKGROUND","FULL","LIMITED"}, userDisallowed = {}})
-    return validate_result
-  end)
+      local validate_result = common_functions_external_consent:ValidateHMIPermissions(data,
+        "SendLocation", {allowed = {"BACKGROUND","FULL","LIMITED"}, userDisallowed = {}})
+      return validate_result
+    end)
+end
+
+--------------------------------------------------------------------------
+-- Precondition:
+-- Register and activate application 2
+--------------------------------------------------------------------------
+RegistAndActivateApp("2", "mobileSession2", config.application2.registerAppInterfaceParams)
+
+--------------------------------------------------------------------------
+-- Main check:
+-- RPC of Group001 is allowed to process on Application 1.
+--------------------------------------------------------------------------
+Test[TEST_NAME_ON .. "MainCheck_RPC_of_Application_1_Group001_is_allowed"] = function(self)
+  local cid = self.mobileSession:SendRPC("SendLocation", {
+      longitudeDegrees = 1.1,
+      latitudeDegrees = 1.1
+    })
+  EXPECT_HMICALL("Navigation.SendLocation")
+  :Do(function(_,data)
+      self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
+    end)
+  self.mobileSession:ExpectResponse("SendLocation", {success = true , resultCode = "SUCCESS"})
 end
 
 --------------------------------------------------------------------------
 -- Main check:
---   RPC is allowed to process.
+-- RPC of Group001 is allowed to process on Application 2.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "MainCheck_RPC_is_allowed"] = function(self)
-	--mobile side: send SubscribeWayPoints request
-  local corid = self.mobileSession:SendRPC("SubscribeWayPoints",{})
-  --hmi side: expected SubscribeWayPoints request
-  EXPECT_HMICALL("Navigation.SubscribeWayPoints")
+Test[TEST_NAME_ON .. "MainCheck_RPC_of_Application_2_Group001_is_allowed"] = function(self)
+  local cid = self.mobileSession2:SendRPC("SendLocation", {
+      longitudeDegrees = 1.1,
+      latitudeDegrees = 1.1
+    })
+  EXPECT_HMICALL("Navigation.SendLocation")
   :Do(function(_,data)
-    --hmi side: sending Navigation.SubscribeWayPoints response
-    self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
-  end)
-  --mobile side: SubscribeWayPoints response
-  EXPECT_RESPONSE("SubscribeWayPoints", {success = true , resultCode = "SUCCESS"})
-  EXPECT_NOTIFICATION("OnHashChange")
+      self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
+    end)
+  self.mobileSession2:ExpectResponse("SendLocation", {success = true , resultCode = "SUCCESS"})
 end
 
--- end Test 05.01
+-- end Test 12.01
 ----------------------------------------------------
 ---------------------------------------------------------------------------------------------
 --------------------------------------Postcondition------------------------------------------
