+++ test_scripts/Polices/External_UCS/External_Consent_ON/029_ATF_Policies_External_Consent_ON_same_notification_user_consent_disallowed_externalConsentStatus_allowed.lua
@@ -1,3 +1,6 @@
+-------------------------------------- Requirement summary -------------------------------------------
+-- [Policies] External UCS: externalConsentStatus vs. consentedFunctions priority
+--
 ------------------------------------------------------------------------------------------------------
 ------------------------------------General Settings for Configuration--------------------------------
 ------------------------------------------------------------------------------------------------------
@@ -6,6 +9,7 @@ local common_functions_external_consent = require('user_modules/shared_testcases
 ------------------------------------------------------------------------------------------------------
 ---------------------------------------Common Variables-----------------------------------------------
 ------------------------------------------------------------------------------------------------------
+local id_group_1
 local policy_file = config.pathToSDL .. "storage/policy.sqlite"
 ------------------------------------------------------------------------------------------------------
 ---------------------------------------Preconditions--------------------------------------------------
@@ -18,33 +22,34 @@ common_steps:ActivateApplication("Activate_Application_1", config.application1.r
 ------------------------------------------Tests-------------------------------------------------------
 ------------------------------------------------------------------------------------------------------
 --------------------------------------------------------------------------
--- TEST 03:
-  -- In case:
-  -- SDL Policies database contains "disallowed_by_external_consent_entities_on" param in "functional grouping" section
-  -- and SDL gets SDL.OnAppPermissionConsent ("externalConsentStatus: ON")
-  -- SDL must sends 'USER_DISALLOWED, success:false' to mobile app on requested RPCs from this "functional grouping"
+-- TEST 04:
+-- In case:
+-- SDL received SDL.OnAppPermissionConsent that contains both (consentedFunctions:allowed, externalConsentStatus)
+-- and according to externalConsentStatus "functional_grouping" for the assigned app is "userDisallowed"
+-- SDL must
+-- change "functional_grouping" status according to externalConsentStatus to "userDisallowed"
 --------------------------------------------------------------------------
--- Test 03.01:
--- Description: disallowed_by_external_consent_entities_on exists. Data consent is disallowed. HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON)
--- Expected Result: requested RPC is disallowed by data consent
+-- Test 04.02:
+-- Description: disallowed_by_external_consent_entities_off exists. HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON, function disallowed)
+-- Expected Result: requested RPC is allowed by External Consent
 --------------------------------------------------------------------------
 -- Precondition:
---   Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
---   Request Policy Table Update.
+-- Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
+-- Request Policy Table Update.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "Precondition_Update_Policy_Table"] = function(self)
+Test[TEST_NAME_ON.."Precondition_Update_Policy_Table"] = function(self)
   -- create json for PTU from sdl_preloaded_pt.json
   local data = common_functions_external_consent:ConvertPreloadedToJson()
   -- insert Group001 into "functional_groupings"
   data.policy_table.functional_groupings.Group001 = {
     user_consent_prompt = "ConsentGroup001",
-    disallowed_by_external_consent_entities_on = {{
-      entityType = 2,
-      entityID = 5
+    disallowed_by_external_consent_entities_off = {{
+        entityType = 2,
+        entityID = 5
     }},
     rpcs = {
-      SubscribeVehicleData = {
-        hmi_levels = {"NONE", "BACKGROUND", "FULL", "LIMITED"}
+      SubscribeWayPoints = {
+        hmi_levels = {"BACKGROUND", "FULL", "LIMITED"}
       }
     }
   }
@@ -59,9 +64,9 @@ Test[TEST_NAME_ON .. "Precondition_Update_Policy_Table"] = function(self)
   --insert "ConsentGroup001" into "consumer_friendly_messages"
   data.policy_table.consumer_friendly_messages.messages["ConsentGroup001"] = {languages = {}}
   data.policy_table.consumer_friendly_messages.messages.ConsentGroup001.languages["en-us"] = {
-        tts = "tts_test",
-        label = "label_test",
-        textBody = "textBody_test"
+    tts = "tts_test",
+    label = "label_test",
+    textBody = "textBody_test"
   }
   -- create json file for Policy Table Update
   common_functions_external_consent:CreateJsonFileForPTU(data, "/tmp/ptu_update.json")
@@ -75,68 +80,64 @@ end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   Check GetListOfPermissions response with empty externalConsentStatus array list. Get group id.
+-- Check GetListOfPermissions response with empty externalConsentStatus array list. Get group id.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "Precondition_GetListOfPermissions"] = function(self)
+Test[TEST_NAME_ON.."Precondition_GetListOfPermissions"] = function(self)
   --hmi side: sending SDL.GetListOfPermissions request to SDL
   local request_id = self.hmiConnection:SendRequest("SDL.GetListOfPermissions")
   -- hmi side: expect SDL.GetListOfPermissions response
   EXPECT_HMIRESPONSE(request_id,{
-    result = {
-      code = 0,
-      method = "SDL.GetListOfPermissions",
-      allowedFunctions = {{name = "ConsentGroup001", allowed = nil}},
-      externalConsentStatus = {}
-    }
-  })
+      result = {
+        code = 0,
+        method = "SDL.GetListOfPermissions",
+        allowedFunctions = {{name = "ConsentGroup001", allowed = nil}},
+        externalConsentStatus = {}
+      }
+    })
+  :Do(function(_,data)
+      id_group_1 = common_functions_external_consent:GetGroupId(data, "ConsentGroup001")
+    end)
 end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   HMI sends OnAllowSDLFunctionality with data consent = disallowed
---------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAllowSDLFunctionality"] = function(self)
-  --hmi side: send request SDL.OnAllowSDLFunctionality
-  self.hmiConnection:SendNotification("SDL.OnAllowSDLFunctionality",
-    {allowed = false, source = "GUI"})
-  self.mobileSession:ExpectNotification("OnPermissionsChange")
-	:ValidIf(function(_,data)
-    local validate_result = common_functions_external_consent:ValidateHMIPermissions(data,
-      "SubscribeVehicleData")
-    return not validate_result
-  end)
-end
-
---------------------------------------------------------------------------
--- Precondition:
---   HMI sends OnAppPermissionConsent with External Consent status = ON
+-- HMI sends OnAppPermissionConsent with consented function = disallowed and External Consent status = ON
 --------------------------------------------------------------------------
 Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent"] = function(self)
+  hmi_app_id_1 = common_functions:GetHmiAppId(config.application1.registerAppInterfaceParams.appName, self)
   -- hmi side: sending SDL.OnAppPermissionConsent for applications
-	self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
-    appID = nil, source = "GUI",
-    externalConsentStatus = {{entityType = 2, entityID = 5, status = "ON"}}
-  })
+  self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
+      appID = hmi_app_id_1, source = "GUI",
+      externalConsentStatus = {{entityType = 2, entityID = 5, status = "ON"}},
+      consentedFunctions = {{name = "ConsentGroup001", id = id_group_1, allowed = false}}
+    })
   self.mobileSession:ExpectNotification("OnPermissionsChange")
-	:ValidIf(function(_,data)
-    local validate_result = common_functions_external_consent:ValidateHMIPermissions(data,
-      "SubscribeVehicleData")
-    return not validate_result
-  end)
+  :ValidIf(function(_,data)
+      local validate_result = common_functions_external_consent:ValidateHMIPermissions(data,
+        "SubscribeWayPoints", {allowed = {"BACKGROUND","FULL","LIMITED"}, userDisallowed = {}})
+      return validate_result
+    end)
 end
 
 --------------------------------------------------------------------------
 -- Main check:
---   RPC is disallowed to process.
+-- RPC is allowed to process.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "MainCheck_RPC_is_disallowed"] = function(self)
-  local corid = self.mobileSession:SendRPC("SubscribeVehicleData",{rpm = true})
-  self.mobileSession:ExpectResponse(corid, {success = false, resultCode = "DISALLOWED"})
+Test[TEST_NAME_ON .. "MainCheck_RPC_is_allowed"] = function(self)
+  --mobile side: send SubscribeWayPoints request
+  local corid = self.mobileSession:SendRPC("SubscribeWayPoints",{})
+  --hmi side: expected SubscribeWayPoints request
+  EXPECT_HMICALL("Navigation.SubscribeWayPoints")
+  :Do(function(_,data)
+      --hmi side: sending Navigation.SubscribeWayPoints response
+      self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
+    end)
+  --mobile side: SubscribeWayPoints response
+  EXPECT_RESPONSE("SubscribeWayPoints", {success = true , resultCode = "SUCCESS"})
   EXPECT_NOTIFICATION("OnHashChange")
-  :Times(0)
 end
 
--- end Test 03.01
+-- end Test 04.01
 ----------------------------------------------------
 ---------------------------------------------------------------------------------------------
 --------------------------------------Postcondition------------------------------------------
