+++ test_scripts/Polices/External_UCS/External_Consent_ON/030_ATF_Policies_External_Consent_ON_empty_consent_groups.lua
@@ -1,3 +1,6 @@
+-------------------------------------- Requirement summary -------------------------------------------
+-- [Policies] External UCS: empty "consent_groups" changes after externalConsentStatus was received
+--
 ------------------------------------------------------------------------------------------------------
 ------------------------------------General Settings for Configuration--------------------------------
 ------------------------------------------------------------------------------------------------------
@@ -11,50 +14,57 @@ local policy_file = config.pathToSDL .. "storage/policy.sqlite"
 ---------------------------------------Preconditions--------------------------------------------------
 ------------------------------------------------------------------------------------------------------
 -- Start SDL and register application
-common_functions_external_consent:PreconditonSteps("mobileConnection","mobileSession")
+common_functions_external_consent:PreconditonSteps("mobileConnection","mobileSession", "mobileSession2")
 -- Activate application
 common_steps:ActivateApplication("Activate_Application_1", config.application1.registerAppInterfaceParams.appName)
-------------------------------------------------------------------------------------------------------
--------------------------------------------Functions--------------------------------------------------
-------------------------------------------------------------------------------------------------------
-local function RegistAndActivateApp(test_case_name, mobile_session, application_params)
-  common_steps:AddMobileSession("Precondition_Add_Mobile_Session_" .. test_case_name, "mobileConnection",mobile_session)
-  common_steps:RegisterApplication("Precondition_Register_Application_" .. test_case_name, mobile_session, application_params)
-  common_steps:ActivateApplication("Precondition_Activate_Application_" .. test_case_name, application_params.appName)
-end
 ------------------------------------------------------------------------------------------------------
 ------------------------------------------Tests-------------------------------------------------------
 ------------------------------------------------------------------------------------------------------
 --------------------------------------------------------------------------
--- TEST 12:
-  -- In case
-  -- SDL received SDL.OnAppPermissionConsent (externalConsentStatus) that changed user's permissions for a "<functional_grouping>"
-  -- and new app that has this "<functional_grouping>" assigned connects
-  -- SDL must
-  -- 1. apply External Consent User Consent Settings to this app
-  -- 2. send corresponding OnPermissionsChange
-  -- 3. add corresponding records to PolicyTable -> "<deviceID>" section
+-- TEST 07:
+-- In case
+-- SDL Policies database does not have "consent_groups" in "device_data" -> "user_consent_records" -> "appID" section
+-- and_ SDL received SDL.OnAppPermissionsConsent (externalConsentStatus)
+-- SDL must
+-- add "consent_groups" to "device_data" -> "user_consent_records" -> "appID" section in Policy Table
+-- change_ according to the received externalConsentStatus settings in both:
+-- "consent_groups" in "device_data" -> "user_consent_records" -> "appID" section
+-- and
+-- "external_consent_status_groups" in "device_data" -> "user_consent_records" -> "appID" section
 --------------------------------------------------------------------------
--- Test 12.01:
--- Description: disallowed_by_external_consent_entities_off. HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON). Register new applications.
--- Expected Result: external_consent is created automatically.
+-- Test 07.01:
+-- Description: "consent_groups" does not exist. disallowed_by_external_consent_entities_on/off. HMI -> SDL: OnAppPermissionConsent(externalConsentStatus ON)
+-- Expected Result: "consent_groups" is created with same status as externalConsentStatus
 --------------------------------------------------------------------------
 -- Precondition:
---   Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
---   Request Policy Table Update.
+-- Prepare JSON file with consent groups. Add all consent group names into app_polices of applications
+-- Request Policy Table Update.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON.."Precondition_Update_Policy_Table_1"] = function(self)
+Test[TEST_NAME_ON.."Precondition_Update_Policy_Table"] = function(self)
   -- create json for PTU from sdl_preloaded_pt.json
   local data = common_functions_external_consent:ConvertPreloadedToJson()
   -- insert Group001 into "functional_groupings"
   data.policy_table.functional_groupings.Group001 = {
     user_consent_prompt = "ConsentGroup001",
     disallowed_by_external_consent_entities_off = {{
-      entityType = 2,
-      entityID = 5
+        entityType = 1,
+        entityID = 4
     }},
     rpcs = {
-      SendLocation = {
+      SubscribeWayPoints = {
+        hmi_levels = {"BACKGROUND", "FULL", "LIMITED"}
+      }
+    }
+  }
+  -- insert Group002 into "functional_groupings"
+  data.policy_table.functional_groupings.Group002 = {
+    user_consent_prompt = "ConsentGroup002",
+    disallowed_by_external_consent_entities_on = {{
+        entityType = 2,
+        entityID = 5
+    }},
+    rpcs = {
+      SubscribeVehicleData = {
         hmi_levels = {"BACKGROUND", "FULL", "LIMITED"}
       }
     }
@@ -73,22 +83,21 @@ Test[TEST_NAME_ON.."Precondition_Update_Policy_Table_1"] = function(self)
     steal_focus = false,
     priority = "NONE",
     default_hmi = "NONE",
-    groups = {"Base-4", "Group001"}
-  }
-  --insert application "0000003" into "app_policies"
-  data.policy_table.app_policies["0000003"] = {
-    keep_context = false,
-    steal_focus = false,
-    priority = "NONE",
-    default_hmi = "NONE",
-    groups = {"Base-4", "Group001"}
+    groups = {"Base-4", "Group001", "Group002"}
   }
   --insert "ConsentGroup001" into "consumer_friendly_messages"
   data.policy_table.consumer_friendly_messages.messages["ConsentGroup001"] = {languages = {}}
   data.policy_table.consumer_friendly_messages.messages.ConsentGroup001.languages["en-us"] = {
-        tts = "tts_test",
-        label = "label_test",
-        textBody = "textBody_test"
+    tts = "tts_test",
+    label = "label_test",
+    textBody = "textBody_test"
+  }
+  --insert "ConsentGroup002" into "consumer_friendly_messages"
+  data.policy_table.consumer_friendly_messages.messages["ConsentGroup002"] = {languages = {}}
+  data.policy_table.consumer_friendly_messages.messages.ConsentGroup002.languages["en-us"] = {
+    tts = "tts_test",
+    label = "label_test",
+    textBody = "textBody_test"
   }
   -- create json file for Policy Table Update
   common_functions_external_consent:CreateJsonFileForPTU(data, "/tmp/ptu_update.json")
@@ -102,82 +111,64 @@ end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   Check GetListOfPermissions response with empty externalConsentStatus array list.
+-- Check GetListOfPermissions response with empty externalConsentStatus array list.
 --------------------------------------------------------------------------
 Test[TEST_NAME_ON.."Precondition_GetListOfPermissions"] = function(self)
-  --hmi side: sending SDL.GetListOfPermissions request to SDL
   local request_id = self.hmiConnection:SendRequest("SDL.GetListOfPermissions")
-  -- hmi side: expect SDL.GetListOfPermissions response
   EXPECT_HMIRESPONSE(request_id,{
-    result = {
-      code = 0,
-      method = "SDL.GetListOfPermissions",
-      allowedFunctions = {
-        {name = "ConsentGroup001", allowed = nil}
-      },
-      externalConsentStatus = {}
-    }
-  })
+      result = {
+        code = 0,
+        method = "SDL.GetListOfPermissions",
+        allowedFunctions = {
+          {name = "ConsentGroup001", allowed = nil},
+          {name = "ConsentGroup002", allowed = nil}
+        },
+        externalConsentStatus = {}
+      }
+    })
 end
 
 --------------------------------------------------------------------------
 -- Precondition:
---   HMI sends OnAppPermissionConsent with External Consent status = ON
+-- HMI sends OnAppPermissionConsent with External Consent status = ON
 --------------------------------------------------------------------------
 Test[TEST_NAME_ON .. "Precondition_HMI_sends_OnAppPermissionConsent"] = function(self)
   hmi_app_id_1 = common_functions:GetHmiAppId(config.application1.registerAppInterfaceParams.appName, self)
   -- hmi side: sending SDL.OnAppPermissionConsent for applications
-	self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
-    appID = hmi_app_id_1, source = "GUI",
-    externalConsentStatus = {{entityType = 2, entityID = 5, status = "ON"}}
-  })
+  self.hmiConnection:SendNotification("SDL.OnAppPermissionConsent", {
+      appID = hmi_app_id_1, source = "GUI",
+      externalConsentStatus = {
+        {entityType = 1, entityID = 4, status = "ON"},
+        {entityType = 2, entityID = 5, status = "ON"}
+      }
+    })
   self.mobileSession:ExpectNotification("OnPermissionsChange")
   :ValidIf(function(_,data)
-    local validate_result = common_functions_external_consent:ValidateHMIPermissions(data,
-      "SendLocation", {allowed = {"BACKGROUND","FULL","LIMITED"}, userDisallowed = {}})
-    return validate_result
-  end)
-end
-
---------------------------------------------------------------------------
--- Precondition:
---   Register and activate application 2
---------------------------------------------------------------------------
-RegistAndActivateApp("2", "mobileSession2", config.application2.registerAppInterfaceParams)
-
---------------------------------------------------------------------------
--- Main check:
---   RPC of Group001 is allowed to process on Application 1.
---------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "MainCheck_RPC_of_Application_1_Group001_is_allowed"] = function(self)
-  local cid = self.mobileSession:SendRPC("SendLocation", {
-    longitudeDegrees = 1.1,
-    latitudeDegrees = 1.1
-  })
-  EXPECT_HMICALL("Navigation.SendLocation")
-  :Do(function(_,data)
-    self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
-  end)
-  self.mobileSession:ExpectResponse("SendLocation", {success = true , resultCode = "SUCCESS"})
+      local validate_result = common_functions_external_consent:ValidateHMIPermissions(data,
+        "SubscribeWayPoints", {allowed = {"BACKGROUND","FULL","LIMITED"}, userDisallowed = {}})
+      return validate_result
+    end)
 end
 
 --------------------------------------------------------------------------
 -- Main check:
---   RPC of Group001 is allowed to process on Application 2.
+-- RPC of Group001 on Application 1 is allowed to process.
 --------------------------------------------------------------------------
-Test[TEST_NAME_ON .. "MainCheck_RPC_of_Application_2_Group001_is_allowed"] = function(self)
-  local cid = self.mobileSession2:SendRPC("SendLocation", {
-    longitudeDegrees = 1.1,
-    latitudeDegrees = 1.1
-  })
-  EXPECT_HMICALL("Navigation.SendLocation")
+Test[TEST_NAME_ON .. "MainCheck_RPC_Of_App1_Group001_is_allowed"] = function(self)
+  --mobile side: send SubscribeWayPoints request
+  local corid = self.mobileSession:SendRPC("SubscribeWayPoints",{})
+  --hmi side: expected SubscribeWayPoints request
+  EXPECT_HMICALL("Navigation.SubscribeWayPoints")
   :Do(function(_,data)
-    self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
-  end)
-  self.mobileSession2:ExpectResponse("SendLocation", {success = true , resultCode = "SUCCESS"})
+      --hmi side: sending Navigation.SubscribeWayPoints response
+      self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS",{})
+    end)
+  --mobile side: SubscribeWayPoints response
+  EXPECT_RESPONSE("SubscribeWayPoints", {success = true , resultCode = "SUCCESS"})
+  EXPECT_NOTIFICATION("OnHashChange")
 end
 
--- end Test 12.01
+-- end Test 07.01
 ----------------------------------------------------
 ---------------------------------------------------------------------------------------------
 --------------------------------------Postcondition------------------------------------------
